consolidating tasks
