--- tasks/lint.js
@@ -1,21 +1,0 @@
-/* eslint-disable */
-
-// TODO: Use stylelint's selector pattern and BEM linter
-
-const gulp       = require('gulp');
-const postcss    = require('gulp-postcss');
-const reporter   = require('postcss-reporter');
-const bemLinter  = require('postcss-bem-linter');
-const stylelint  = require('stylelint');
-const paths      = require('../paths');
-
-const linters = [
-  stylelint(),
-  bemLinter('bem'),
-  reporter({ clearMessages: true }),
-];
-
-module.exports = () => {
-  return gulp.src([paths.cssSrc, paths.cssLib])
-    .pipe(postcss(linters));
-};
